requirements.txt
Django>=4.2.5
django-cleanup>=8.0.0
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
typing_extensions>=4.12.2
drf-yasg>=1.21.7